chore(NODE-3717): test reorg part a-c (#3097)
