+++ test/integration/node-specific/bson-options/bsonRegExp.test.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const { expect } = require('chai');
-const { BSONRegExp } = require('../../src/index');
+const { BSONRegExp } = require('../../../../src/index');
 
 describe('BSONRegExp', () => {
   describe('bsonRegExp option', () => {
